Improved test scripts structure
