+++ tests/test_analysis_methods.py
@@ -7,10 +7,9 @@ import shutil
 import unittest
 import os
 
-print sys.path
+from transit_test import *
 
 import pytransit
-print pytransit.__path__
 import pytransit.norm_tools as norm_tools
 
 from pytransit.analysis.gumbel import GumbelMethod
@@ -22,39 +21,8 @@ from pytransit.analysis.resampling import ResamplingMethod
 from pytransit.analysis.rankproduct import RankProductMethod
 
 
-ctrl_rep1 = "../src/pytransit/data/glycerol_H37Rv_rep1.wig"
-ctrl_rep2 = "../src/pytransit/data/glycerol_H37Rv_rep2.wig"
-ctrl_data_txt = ",".join([ctrl_rep1, ctrl_rep2])
-
-exp_rep1 = "../src/pytransit/data/cholesterol_H37Rv_rep1.wig"
-exp_rep2 = "../src/pytransit/data/cholesterol_H37Rv_rep2.wig"
-exp_rep3 = "../src/pytransit/data/cholesterol_H37Rv_rep3.wig"
-exp_data_txt = ",".join([exp_rep1, exp_rep2, exp_rep3])
-
-annotation = "../src/pytransit/genomes/H37Rv.prot_table"
-small_annotation = "test.prot_table"
-output = "testoutput.txt"
-
-class TestMethods(unittest.TestCase):
+class TestMethods(TransitTestCase):
  
-    def setUp(self):
-
-        print ""
-        print "#"*20
-        if os.path.exists(output):
-            print "Removing output file..."
-            os.remove(output) 
-
-        hist_path = output.rsplit(".", 1)[0] + "_histograms"
-        if os.path.exists(hist_path):
-            print "Removing histogram dir..."
-            shutil.rmtree(hist_path)
-    
-        genes_path = output.rsplit(".", 1)[0] + "_genes" + output.rsplit(".", 1)[1]
-        if os.path.exists(genes_path):
-            print "Removing genes file..."
-            os.remove(genes_path)
-
  
     def test_Gumbel(self):
         args = [ctrl_data_txt, annotation, output, "-s", "1000", "-b", "100"]
