+++ tests/test_pytransit_tools.py
@@ -1,50 +1,22 @@
 import sys
-
 sys.path.insert(0, '../src/')
 
+import os
 import shutil
 import unittest
 import os
 import numpy
+
+from transit_test import *
 
 import pytransit.norm_tools as norm_tools
 import pytransit.tnseq_tools as tnseq_tools
 import pytransit.stat_tools as stat_tools
 
 
-ctrl_rep1 = "../src/pytransit/data/glycerol_H37Rv_rep1.wig"
-ctrl_rep2 = "../src/pytransit/data/glycerol_H37Rv_rep2.wig"
-ctrl_data_txt = ",".join([ctrl_rep1, ctrl_rep2])
 
-exp_rep1 = "../src/pytransit/data/cholesterol_H37Rv_rep1.wig"
-exp_rep2 = "../src/pytransit/data/cholesterol_H37Rv_rep2.wig"
-exp_rep3 = "../src/pytransit/data/cholesterol_H37Rv_rep3.wig"
-exp_data_txt = ",".join([exp_rep1, exp_rep2, exp_rep3])
-
-all_data_list = [ctrl_rep1, ctrl_rep2, exp_rep1, exp_rep2, exp_rep3]
-
-annotation = "../src/pytransit/genomes/H37Rv.prot_table"
-
-output = "testoutput.txt"
-
-
-class TestTnSeqTools(unittest.TestCase):
+class TestTnSeqTools(TransitTestCase):
  
-    def setUp(self):
-
-        print "#"*20
-        print self.id()
-        print "#"*20
-        print "\n"
-        if os.path.exists(output):
-            print "Removing output file..."
-            os.remove(output) 
-    
-        genes_path = output.rsplit(".", 1)[0] + "_genes" + output.rsplit(".", 1)[1]
-        if os.path.exists(genes_path):
-            print "Removing genes file..."
-            os.remove(genes_path)
-
 
     def test_read_data(self):
         data,position = tnseq_tools.get_data(all_data_list)
